simplify tests folder structure
